@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,12 +34,12 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  animated_bottom_navigation_bar: ^1.2.0
-  page_transition: ^2.0.9
-  mobile_scanner: ^3.1.0
-  sqflite: ^2.2.6
-  firebase_auth: ^4.3.0
+  cupertino_icons: ^1.0.8
+  animated_bottom_navigation_bar: ^1.3.3
+  page_transition: ^2.1.0
+  mobile_scanner: ^7.0.1
+  sqflite: ^2.3.3
+  firebase_auth: ^5.3.1
 
 
 
@@ -47,7 +47,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.12.0
+  flutter_launcher_icons: ^0.13.1
 
 flutter_icons:
   android: true
@@ -60,7 +60,7 @@ flutter_icons:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
