@@ -8,7 +8,7 @@ import Foundation
 import firebase_auth
 import firebase_core
 import mobile_scanner
-import sqflite
+import sqflite_darwin
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   FLTFirebaseAuthPlugin.register(with: registry.registrar(forPlugin: "FLTFirebaseAuthPlugin"))
